Added base line code for adding to local storage, needs finish hooking up reducer
