+++ src/redux/reducers/index.js
@@ -3,4 +3,5 @@ import { combineReducers } from "redux";
 
 export default combineReducers({
     //list reducers
+    userData
 });
